requirements.txt
django
djangorestframework
djangorestframework-simplejwt
faker